requirements.txt
wholeslidedata == 0.0.16
numpy >= 1.20.2
Pillow == 9.1.0
PyYAML == 6.0
Shapely == 1.7.1
albumentations == 1.1.0
cmapy == 0.6.6
dropblock == 0.3.0
h5py == 3.6.0
lifelines == 0.27.0
lightning_bolts == 0.6.0.post1
lxml == 4.8.0
matplotlib == 3.6.2
opencv_python_headless == 4.5.4.60
openslide_python == 1.1.2
pandas == 1.4.2
pycm == 3.4
pycox == 0.2.3
pynvml == 8.0.4
pytorch_lightning == 1.7.0
ray == 1.11.0
scikit_image == 0.19.2
scikit_learn == 1.0.2
scipy == 1.8.0
tabulate == 0.8.9
timm == 0.6.12
torch == 1.10.0
torchmetrics == 0.7.3
torchvision == 0.11.1
tqdm == 4.64.0
monai == 0.9.0